refactoring and extracting components
