+++ src/components/App.js
@@ -2,7 +2,7 @@ import React, { useEffect, useState } from "react";
 import { get } from "utils/requests";
 import Titlebar from "components/titlebar/Titlebar";
 import styles from "components/App.module.scss";
-import MainComponent from "./main/MainComponent";
+import MainComponent from "./MainComponent";
 
 function App() {
   const [response, setResponse] = useState([]);
