--- src/components/main/Main.module.scss
@@ -1,44 +1,0 @@
-.main_container {
-  text-align: center;
-  display: flex;
-  flex-direction: row;
-  height: 90vh;
-  /* background-color: blue; */
-  /* border-color: blue; */
-  border-radius: 10px;
-  flex: 6;
-  padding: 4px
-}
-
-.sidebar {
-  display: flex;
-  flex-direction: column;
-  height: 100%;
-  /* background-color: green;*/
-
-
-  flex: 2;
-  justify-content: space-between;
-}
-
-.sidebar_top_container {
-  display: flex;
-  align-self: flex-start;
-  justify-content: space-between;
-  /* background-color: brown;*/
-}
-
-.bottom_bar{
-  display: flex;
-  justify-content: space-around;
-  /* background-color: orange; */
-  align-self: flex-end;
-}
-
-.tab_box{
-  /* background-color: pink; */
-  display: flex;
-  flex: 4;
-  height: 100%;
-
-}
